requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.32.0
azure-identity>=1.19.0
python-dotenv>=1.0.1
azure-monitor-opentelemetry>=1.6.0
opentelemetry-instrumentation-fastapi>=0.50b0
azure-servicebus>=7.12.0
azure-storage-blob>=12.22.0
azure-ai-evaluation>=0.4.0
azure-ai-agents>=1.0.0
azure-ai-projects>=1.0.0
opentelemetry-api>=1.27.0
